requirements.txt
gunicorn==21.2.0
flask==2.3.3
google-generativeai==0.3.1
Pillow==10.0.0
